DialogGraphNodes can req for cons to be removed (#29)
This PR implements the `remove_connections_request` signal to DialogGraphNode. This signal allows dialog graph nodes to request that connections in the graph be removed on specific connection slots of node. This is necessary for things like the ChoicePrompt node, which will need to request that slots are disconnected when removed.