Dialog Graph Refactor - Packed & Runtime Graphs (#32)
This PR totally refactors how dialog graphs are stored and constructed. The result is DialogGraph being split into two structures - PackedDialogGraph and RuntimeDialogGraph.

This has been done in order to make DialogGraphs both less and more reference-based. When packed, DialogGraphs boil down to the data that the represent as a series of dictionaries. When unpacked, DialogGraphs inflate to a structure optimized for runtime use. This eliminates the use of Resources for most of the graph data, in order to address fixing #24. It simplifies a lot of code for working with dialog graphs, but adds a little overhead to how data is stored for each graph.