Item Plugin First Draft (#35)
This PR creates a first pass of the Item system. It implements basic items and inventories in a way intended to be flexible enough for multiple kinds of games. The source engine & crafting-based games were large inspirations while working on this.

Types of items are defined using an `ItemDescriptor` resource, which allows one to provide the name, icon, view-model, and world entity for an item. ItemDescriptors function as a sort of factory for producing that specific type of item, and produce `ItemInstance`s. These represent an instance of some item (or stack of items), and can be freely moved between inventories and the game world.

This plugin is still missing key functionality for item interactions, viewmodels, inventory filters, and inventory serialization. Those will come in following PRs.