Resource-based Item Filtering (#48)
This PR implements item filtering through the use of a new resource named `ItemFilter`. The `ItemFilter` class is fairly simple, and just defines an interface for subclasses to implement filtering logic. A handful of filters have been implemented by default, allowing one to filter by the category of an item, by the name or display name of an item, by the type of item, and compound these filters using AND / OR filters as well.

Inventories support filters, and now have an exposed field for a filter to be used when items are added to said inventory.