Refactored code to work with a RayCast3D node and followed other requests
Now the code generates a RayCast3D node instead of completely calculating a ray cast which I did it that way so that it can still be modular instead of having a permanent node on the player controller and this actually ended up simplifying the code further which I wasn't completely expecting. I also add with the minor offset now instead of subtracting in order to automatically detect if a player is walking into an space that has a surface just barely above them so they automatically crouch. For some reason when implementing the RayCast3D node I had to take a 0 out of the minor offset so that's why it's slightly bigger now. I also changed the code so that way the origin of the ray cast is always at the player's position as requested. This is at the point now where I'm unsure if it can be refined further but I guess we'll see, I mean the equations for the target position of the ray cast just seem to be getting simpler and simpler which I wasn't really expecting when changing the code to work with a RayCast3D node.