Fully Implemented Saving and Loading the Knowledge Junction Node
It took me a while but I eventually figured it out and now we can fully save and load all settings with the knowledge junction node. Next up is now actually going to be it's unique functionality because there is actually nothing else to do for it besides clean up the code I've written. I also as a byproduct ended up removing basically all the code that was initially there to work for the choice prompt node so that's nice too. I also found and fixed a bug with the choice prompt node where the text for each prompt wasn't saving properly which was because it was being assigned to a variable that was entirely unused when loading it instead of the key that was intended to be used when loading it. Also I now see why the Knowledge Action node wasn't fully included in the graph yet as it is not setup to actually store data yet nor perform any actually unique functionality, however it was great practice for adding nodes to the graph for me.