Fully Implemented Functionality of Knowledge Junction Node
I finally got around to doing it, and it was shockingly easier than implementing saving and loading with the node. I also made it so the field for putting in the constant value allowed for floats and I lastly attempted to comment everything with the notation seen within the rest of netengine5.