Removed changes made to Knowledge Action
Since Knowledge Action is not fully implemented and the point of this pull request is only to implement the Knowledge Junction mode I decided to undo the changes I made to allow Knowledge Action to appear in the graph. However, I will keep the bug fix I made for Choice Prompt as I feel that while it wasn't the main focus of this pull request it would be a nice addition to it.