Refactor `PlayerController` out to `CharacterAgent3D` (#78)
Takes re-usable functionality out of `PlayerController` and puts it into a new node, `CharacterAgent3D`. The purpose of a char agent is to be some in-game representation of a character. Functionality has been separated in prep for drafting the NPC system, which will benefit from re-using a lot of player code.