`VirtualCamera`: Priorities! (#96)
Implements priorities into `VirtualCamera2D` and `VirtualCamera3D`. These work by having a main priority, which is a general determination of the vcams priority against other vcams. There is also specific priorities, which assign a priority to a vcam on a per-viewport basis. This is so we can use an `EntityCamera` in different viewports, and be able to have different cams have different priorities in the same world!