MouseLook Character Script uses EntityInput (#111)
Refactors the MouseLook CharacterScript to use it's CharacterAgent's EntityInput for mouse input. To do this, the spec for analog inputs have been revised to NOT be between 0 and 1, or -1 and 1. They're now uncapped to account for rapid mouse movements.

Because this takes advantage of EntityInput's analog values, this means mouselook can now be simulated with SimulatedInput for NPC Agents.