`PlayerModel` / Renderer Animation Refactor (#135)
Refactors the `PlayerModel` node into two nodes - `PlayerModel` and `PlayerModelRenderer` (which both have a 2D and 3D variant. `PlayerModelRenderer` takes on the role that the old playermodel node had - acting as a way to establish spawning / swapping playermodels. The new `PlayerModel` node now takes on the role of managing the actual playermodel itself, and should always be the root of a PlayerModel scene. This node contains a reference to the AnimationTree and gracefully handle passing settings into animation params and such.

This PR effectively allows you to animate playermodels from an agent's behaviors. yay!